requirements.txt
mss==9.0.1
Flask==3.0.0
Pillow==10.1.0
PyYAML==6.0.1
ImageHash==4.3.1